Release new gem as 'sinatra-export' and use Sinatra namespace
